Fix netstandard and netcoreapp build issues
