--- src/Common/src/CoreLib/System/IO/PathInternal.IsRoot.cs
@@ -1,16 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Diagnostics;
-using System.Text;
-
-namespace System.IO
-{
-    /// <summary>Contains internal path helpers that are shared between many projects.</summary>
-    internal static partial class PathInternal
-    {
-        internal static bool IsRoot(ReadOnlySpan<char> path)
-            => path.Length == GetRootLength(path);
-    }
-}
