+++ src/System.IO.FileSystem.AccessControl/src/System.IO.FileSystem.AccessControl.csproj
@@ -1,17 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <DefineConstants Condition="'$(TargetGroup)' == 'netstandard' OR '$(TargetsNetFx)' == 'true'">$(DefineConstants);INTERNAL_NULLABLE_ATTRIBUTES</DefineConstants>
+    <DefineConstants Condition="'$(TargetsNetCoreApp)' != 'true'">$(DefineConstants);INTERNAL_NULLABLE_ATTRIBUTES</DefineConstants>
     <AllowUnsafeBlocks Condition="'$(TargetsWindows)' == 'true'">true</AllowUnsafeBlocks>
     <IsPartialFacadeAssembly Condition="'$(TargetsNetFx)'=='true'">true</IsPartialFacadeAssembly>
     <GeneratePlatformNotSupportedAssemblyMessage Condition="'$(TargetsWindows)' != 'true'">SR.PlatformNotSupported_AccessControl</GeneratePlatformNotSupportedAssemblyMessage>
     <Configurations>net461-Windows_NT-Debug;net461-Windows_NT-Release;netcoreapp-Windows_NT-Debug;netcoreapp-Windows_NT-Release;netfx-Windows_NT-Debug;netfx-Windows_NT-Release;netstandard-Debug;netstandard-Release;netstandard-Windows_NT-Debug;netstandard-Windows_NT-Release</Configurations>
   </PropertyGroup>
-  <!-- NetFX/NetStandard for handling nullability in files imported from other assemblies -->
-  <ItemGroup Condition="'$(TargetGroup)' == 'netstandard' OR '$(TargetsNetFx)' == 'true'">
+  <!-- Handle nullability in NetFX -->
+  <ItemGroup Condition="'$(TargetsNetCoreApp)' != 'true'">
     <Compile Include="$(CommonPath)\CoreLib\System\Diagnostics\CodeAnalysis\NullableAttributes.cs" Link="System\Diagnostics\CodeAnalysis\NullableAttributes.cs" />
   </ItemGroup>
-  <!-- Windows (NOT NetFX) files -->
-  <ItemGroup Condition="'$(TargetsWindows)' == 'true' AND '$(TargetsNetFx)' != 'true'">
+  <!-- Source includes -->
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' and '$(TargetsNetFx)' != 'true'">
+    <Compile Include="$(CommonPath)\Interop\Windows\Interop.Errors.cs" Link="Common\Interop\Windows\Interop.Errors.cs" />
+    <Compile Include="$(CommonPath)\System\NotImplemented.cs" Link="Common\System\NotImplemented.cs" />
+    <Compile Include="System\IO\FileSystemAclExtensions.cs" />
+    <Compile Include="System\Security\AccessControl\DirectoryObjectSecurity.cs" />
+    <Compile Include="System\Security\AccessControl\DirectorySecurity.cs" />
+    <Compile Include="System\Security\AccessControl\FileSecurity.cs" />
+    <Compile Include="System\Security\AccessControl\FileSystemAccessRule.cs" />
+    <Compile Include="System\Security\AccessControl\FileSystemAuditRule.cs" />
+    <Compile Include="System\Security\AccessControl\FileSystemRights.cs" />
+    <Compile Include="System\Security\AccessControl\FileSystemSecurity.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Interop.BOOL.cs" Link="Common\CoreLib\Interop\Windows\Interop.BOOL.cs" />
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.FILE_TIME.cs" Link="Common\CoreLib\Interop\Windows\Interop.FILE_TIME.cs" />
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.FileAttributes.cs" Link="Common\Interop\Windows\Interop.FileAttributes.cs" />
@@ -29,14 +41,10 @@
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.WIN32_FILE_ATTRIBUTE_DATA.cs" Link="Common\CoreLib\Interop\Windows\Interop.WIN32_FILE_ATTRIBUTE_DATA.cs" />
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.WIN32_FIND_DATA.cs" Link="Common\CoreLib\Interop\Windows\Interop.WIN32_FIND_DATA.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\DisableMediaInsertionPrompt.cs" Link="Common\System\IO\DisableMediaInsertionPrompt.cs" />
-    <Compile Include="$(CommonPath)\CoreLib\System\IO\Path.DirectorySeparator.cs" Link="Common\CoreLib\System\IO\Path.DirectorySeparator.cs" />
-    <Compile Include="$(CommonPath)\CoreLib\System\IO\Path.Root.Windows.cs" Link="Common\CoreLib\System\IO\Path.Root.Windows.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\PathHelper.Windows.cs" Link="Common\CoreLib\System\IO\PathHelper.Windows.cs" />
-    <Compile Include="$(CommonPath)\CoreLib\System\IO\PathInternal.IsRoot.cs" Link="Common\CoreLib\System\IO\PathInternal.IsRoot.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\PathInternal.Windows.cs" Link="Common\CoreLib\System\IO\PathInternal.Windows.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\Win32Marshal.cs" Link="Common\CoreLib\System\IO\Win32Marshal.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\Text\ValueStringBuilder.cs" Link="Common\CoreLib\System\Text\ValueStringBuilder.cs" />
-    <Compile Include="$(CommonPath)\Interop\Windows\Interop.Errors.cs" Link="Common\Interop\Windows\Interop.Errors.cs" />
     <Compile Include="$(CommonPath)\Interop\Windows\Interop.Libraries.cs" Link="Common\Interop\Windows\Interop.Libraries.cs" />
     <Compile Include="$(CommonPath)\Interop\Windows\Interop.LongFileTime.cs" Link="Common\Interop\Windows\Interop.LongFileTime.cs" />
     <Compile Include="$(CommonPath)\Interop\Windows\Kernel32\Interop.CreateDirectory.cs" Link="Common\Interop\Windows\Interop.CreateDirectory.cs" />
@@ -44,35 +52,27 @@
     <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeFindHandle.Windows.cs" Link="Common\Microsoft\Win32\SafeHandles\SafeFindHandle.Windows.cs" />
     <Compile Include="$(CommonPath)\System\IO\FileSystem.Attributes.Windows.cs" Link="Common\System\IO\FileSystem.Attributes.Windows.cs" />
     <Compile Include="$(CommonPath)\System\IO\FileSystem.DirectoryCreation.Windows.cs" Link="Common\System\IO\FileSystem.DirectoryCreation.Windows.cs" />
-    <Compile Include="$(CommonPath)\System\NotImplemented.cs" Link="Common\System\NotImplemented.cs" />
-    <Compile Include="System\IO\FileSystemAclExtensions.cs" />
-    <Compile Include="System\Security\AccessControl\DirectoryObjectSecurity.cs" />
-    <Compile Include="System\Security\AccessControl\DirectorySecurity.cs" />
-    <Compile Include="System\Security\AccessControl\FileSecurity.cs" />
-    <Compile Include="System\Security\AccessControl\FileSystemAccessRule.cs" />
-    <Compile Include="System\Security\AccessControl\FileSystemAuditRule.cs" />
-    <Compile Include="System\Security\AccessControl\FileSystemRights.cs" />
-    <Compile Include="System\Security\AccessControl\FileSystemSecurity.cs" />
+    <Compile Include="System\IO\FileSystemAclExtensions.netcoreapp.cs" />
   </ItemGroup>
-  <!-- Windows (NOT NetFX) references -->
-  <ItemGroup Condition="'$(TargetsWindows)' == 'true' AND '$(TargetsNetFx)' != 'true'">
+  <ItemGroup Condition="'$(TargetsNetFx)' == 'true'">
+    <Compile Include="System\IO\FileSystemAclExtensions.net46.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' and '$(TargetGroup)' == 'netstandard'">
+    <Compile Include="System\IO\FileSystemAclExtensions.netstandard.cs" />
+  </ItemGroup>
+  <!-- Reference includes -->
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' and '$(TargetsNetFx)' != 'true'">
     <Reference Include="System.Buffers" />
     <Reference Include="System.Collections" />
     <Reference Include="System.Memory" />
     <Reference Include="System.Runtime.InteropServices" />
     <Reference Include="System.Threading.Tasks" />
   </ItemGroup>
-  <!-- NetFX files -->
-  <ItemGroup Condition="'$(TargetsNetFx)'=='true'">
-    <Compile Include="System\IO\FileSystemAclExtensions.net46.cs" />
-  </ItemGroup>
-  <!-- NetFX references -->
-  <ItemGroup Condition="'$(TargetsNetFx)'=='true'">
+  <ItemGroup Condition="'$(TargetsNetFx)' == 'true'">
     <Reference Include="mscorlib" />
     <Reference Include="System.Runtime" />
   </ItemGroup>
-  <!-- NOT NetFX references -->
-  <ItemGroup Condition="'$(TargetsNetFx)'!='true'">
+  <ItemGroup Condition="'$(TargetsNetFx)' != 'true'">
     <Reference Include="System.Collections.NonGeneric" />
     <Reference Include="System.Diagnostics.Debug" />
     <Reference Include="System.IO.FileSystem" />
