+++ src/System.IO.FileSystem/src/System.IO.FileSystem.csproj
@@ -16,7 +16,6 @@
   <!-- All targets files -->
   <ItemGroup>
     <Compile Include="$(CommonPath)\CoreLib\System\IO\PathInternal.cs" Link="Common\CoreLib\System\IO\PathInternal.cs" />
-    <Compile Include="$(CommonPath)\CoreLib\System\IO\PathInternal.IsRoot.cs" Link="Common\CoreLib\System\IO\PathInternal.IsRoot.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\StreamHelpers.CopyValidation.cs" Link="Common\CoreLib\System\IO\StreamHelpers.CopyValidation.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\Text\ValueStringBuilder.cs" Link="Common\CoreLib\System\Text\ValueStringBuilder.cs" />
     <Compile Include="$(CommonPath)\System\IO\PathInternal.CaseSensitivity.cs" Link="Common\System\IO\PathInternal.CaseSensitivity.cs" />
